requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
pydantic==2.8.2
pydantic_core==2.20.1
uvicorn==0.30.1
sqlmodel==0.0.19
pydantic-settings==2.3.4
psycopg==3.2.1
psycopg-binary==3.2.1
python-dotenv==1.0.1
requests==2.32.3
